Add a script to switch mySociety deployments to using shared files
This is required by our switch to timestamped-based deploys.  There
are certain files and directories (e.g. cache, the Xapian databases,
generated graphs) that should be shared between deployments,
rather than recreating them when creating the next deployment in
a new directory.

This script should be run as a one-off before switching an
instance to using timestamped deploys, to ensure that the existing
data is stored in the shared subdirectory of the vhost directory,
so that any future deployment can simply create symlinks into the
shared directory.